+++ src/emmy/viewer/sci.cljs
@@ -2,6 +2,7 @@
   (:require [demo.mathbox]
             [demo.mathlive]
             [emmy.mathbox.plot]
+            [emmy.viewer.plot]
             [leva.sci]
             [emmy.sci]
             [mafs.sci]
@@ -23,4 +24,5 @@
    {:namespaces
     {'demo.mathbox      (sci/copy-ns demo.mathbox (sci/create-ns 'demo.mathbox))
      'demo.mathlive     (sci/copy-ns demo.mathlive (sci/create-ns 'demo.mathlive))
-     'emmy.mathbox.plot (sci/copy-ns emmy.mathbox.plot (sci/create-ns 'emmy.mathbox.plot))}}))
+     'emmy.mathbox.plot (sci/copy-ns emmy.mathbox.plot (sci/create-ns 'emmy.mathbox.plot))
+     'emmy.viewer.plot  (sci/copy-ns emmy.viewer.plot (sci/create-ns 'emmy.viewer.plot))}}))